restructures PostNewCampsite screen to render a MultiStepPostNewCampsiteForm,restructuring elements into a sub folder, and creating a FormStep and NewCampsiteRequestData states to manage multiple components
