+++ parkfinite2/components/map-drawer/post-new-campsite-tab/post-new-campsite-components/form-components/google-places-mini-component.tsx
